feat: Add GuEcsTask construct for non-scheduled tasks
